// src/app/modules/period/period-error.enum.ts
export const enum PeriodErrorEnum {
  EMPTY_SPLIT_NAME = 'Split name is required',
  EMPTY_SPLIT_DATE = 'Split date is required',
  EMPTY_SPLIT_DURATION = 'Split duration is required',
  EMPTY_SPLIT_SET = 'Split is required',
  EMPTY_PERIOD_NAME = 'Period name is required',
  EMPTY_PERIOD_DATE_FROM = 'Period start date is required',
  EMPTY_PERIOD_DATE_DO = 'Period end date is required',
}